+++ test/peddler/notifications/b2b_any_offer_changed_test.rb
@@ -7,17 +7,17 @@ require "helper"
 
 module Peddler
   module Notifications
-    class FeedProcessingFinishedTest < Minitest::Test
+    class B2BAnyOfferChangedTest < Minitest::Test
       include FixtureHelpers
 
       def setup
-        @payload = load_fixture("notifications/feed_processing_finished.json")
+        @payload = load_fixture("notifications/b2b_any_offer_changed.json")
       end
 
       def test_parses_notification
-        notification = FeedProcessingFinished::Notification.parse(@payload)
+        notification = B2BAnyOfferChanged.parse(@payload)
 
-        assert_equal("FEED_PROCESSING_FINISHED", notification.notification_type)
+        assert_equal("B2B_ANY_OFFER_CHANGED", notification.notification_type)
       end
     end
   end