+++ test/peddler/notifications/branded_item_content_change_test.rb
@@ -7,17 +7,17 @@ require "helper"
 
 module Peddler
   module Notifications
-    class FeedProcessingFinishedTest < Minitest::Test
+    class BrandedItemContentChangeTest < Minitest::Test
       include FixtureHelpers
 
       def setup
-        @payload = load_fixture("notifications/feed_processing_finished.json")
+        @payload = load_fixture("notifications/branded_item_content_change.json")
       end
 
       def test_parses_notification
-        notification = FeedProcessingFinished::Notification.parse(@payload)
+        notification = BrandedItemContentChange.parse(@payload)
 
-        assert_equal("FEED_PROCESSING_FINISHED", notification.notification_type)
+        assert_equal("BRANDED_ITEM_CONTENT_CHANGE", notification.notification_type)
       end
     end
   end