+++ test/peddler/notifications/fba_inventory_availability_change_test.rb
@@ -7,17 +7,17 @@ require "helper"
 
 module Peddler
   module Notifications
-    class FeedProcessingFinishedTest < Minitest::Test
+    class FBAInventoryAvailabilityChangeTest < Minitest::Test
       include FixtureHelpers
 
       def setup
-        @payload = load_fixture("notifications/feed_processing_finished.json")
+        @payload = load_fixture("notifications/fba_inventory_availability_change.json")
       end
 
       def test_parses_notification
-        notification = FeedProcessingFinished::Notification.parse(@payload)
+        notification = FBAInventoryAvailabilityChange.parse(@payload)
 
-        assert_equal("FEED_PROCESSING_FINISHED", notification.notification_type)
+        assert_equal("FBA_INVENTORY_AVAILABILITY_CHANGES", notification.notification_type)
       end
     end
   end