+++ test/peddler/notifications/fba_outbound_shipment_status_test.rb
@@ -7,17 +7,17 @@ require "helper"
 
 module Peddler
   module Notifications
-    class FeedProcessingFinishedTest < Minitest::Test
+    class FBAOutboundShipmentStatusTest < Minitest::Test
       include FixtureHelpers
 
       def setup
-        @payload = load_fixture("notifications/feed_processing_finished.json")
+        @payload = load_fixture("notifications/fba_outbound_shipment_status.json")
       end
 
       def test_parses_notification
-        notification = FeedProcessingFinished::Notification.parse(@payload)
+        notification = FBAOutboundShipmentStatus.parse(@payload)
 
-        assert_equal("FEED_PROCESSING_FINISHED", notification.notification_type)
+        assert_equal("FBA_OUTBOUND_SHIPMENT_STATUS", notification.notification_type)
       end
     end
   end